Exchanged old and new functions
