+++ net2brain/architectures/implemented_models/cornet_rt_model.py
