+++ net2brain/architectures/implemented_models/cornet_s_model.py
