+++ net2brain/architectures/implemented_models/cornet_z_model.py
