+++ net2brain/architectures/netsetbase.py
@@ -1,12 +1,11 @@
-
-from torchvision.models.feature_extraction import get_graph_node_names, create_feature_extractor
-from .shared_functions import imagenet_preprocess, imagenet_preprocess_frames, torch_clean, load_from_json
+from torchvision import transforms as trn
 import torchextractor as tx
 from PIL import Image
 import cv2
 import librosa
 import torch
 import torch.nn as nn
+import warnings
 
 # Base class for all NetSets
 class NetSetBase:
@@ -40,11 +39,18 @@ class NetSetBase:
     
     def select_model_layers(self, layers_to_extract, network_layers, loaded_model):
         if layers_to_extract:
-            return layers_to_extract
+            available_layers = tx.list_module_names(loaded_model)
+            valid_layers = [layer for layer in layers_to_extract if layer in available_layers and layer != '']
+            invalid_layers = set(layers_to_extract) - set(valid_layers)
+            if invalid_layers:
+                warnings.warn(f"Some layers are not present in the model and will not be extracted: {invalid_layers}. "
+                            "Please call the 'layers_to_extract()' function from the FeatureExtractor to see all available layers.")
+            return valid_layers
         elif network_layers:
-            return network_layers
+            return [layer for layer in network_layers if layer != '']
         else:
-            return tx.list_module_names(loaded_model)
+            return [layer for layer in tx.list_module_names(loaded_model) if layer != '']
+
 
 
     def __init_subclass__(cls, **kwargs):
@@ -74,10 +80,33 @@ class NetSetBase:
         raise NotImplementedError
 
     def image_preprocessing(self, image, model_name, device):
-        raise NotImplementedError
+        """
+        Args:
+            image (Union[Image.Image, List[Image.Image]]): A PIL Image or a list of PIL Images.
+            model_name (str): The name of the model, used to determine specific preprocessing if necessary.
+            device (str): The device to which the tensor should be transferred ('cuda' for GPU, 'cpu' for CPU).
+
+        Returns:
+            Union[torch.Tensor, List[torch.Tensor]]: The preprocessed image(s) as PyTorch tensor(s).
+        """
+        transforms = trn.Compose([
+            trn.Resize((224, 224)),
+            trn.ToTensor(),
+            trn.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
+        ])
+
+        img_tensor = transforms(image).unsqueeze(0)
+        if device == 'cuda':
+            img_tensor = img_tensor.cuda()
+        return img_tensor
+
 
     def video_preprocessing(self, frame, model_name, device):
-        raise NotImplementedError
+        # Convert numpy array to PIL Image
+        pil_frame = Image.fromarray(frame)
+        
+        return self.image_preprocessing(pil_frame, model_name, device)
+
 
     def clean_extracted_features(self, features):
         # return features
@@ -96,8 +125,47 @@ class NetSetBase:
         raise NotImplementedError
     
     
+    def combine_image_data(self, feature_list):
+        return feature_list[0]
+    
+    def combine_video_data(self, feature_list):
+        """
+        Averages the features extracted from multiple frames of a video.
+
+        Args:
+            feature_list (List[Dict[str, torch.Tensor]]): A list where each element is a dictionary. The keys of the
+            dictionary are the layer names, and the values are the feature tensors extracted from a frame.
+
+        Returns:
+            Dict[str, torch.Tensor]: A dictionary where the keys are the layer names, and the values are the averaged 
+            feature tensors across all frames.
+        """
+        # Initialize a dictionary to store the sum of features for each layer
+        summed_features = {}
+
+        for features in feature_list:
+            for layer, data in features.items():
+                if layer not in summed_features:
+                    # If the layer is not in the dictionary, add it
+                    summed_features[layer] = data.clone()
+                else:
+                    # If the layer is already in the dictionary, accumulate the features
+                    summed_features[layer] += data
+
+        # Calculate the average for each layer
+        num_frames = len(feature_list)
+        averaged_features = {layer: data / num_frames for layer, data in summed_features.items()}
+
+        return averaged_features
+    
+
+    def combine_audio_data(self, feature_list):
+        raise NotImplementedError
+    
+
+
     def load_image_data(self, data_path):
-        return Image.open(data_path).convert('RGB')
+        return [Image.open(data_path).convert('RGB')]
     
     def load_video_data(self, data_path):
         # Logic to load video data using cv2
@@ -119,7 +187,9 @@ class NetSetBase:
         return y, sr
     
 
+
     def randomize_weights(self, m):
+        warnings.warn("Will initiate random weights")
         if isinstance(m, nn.Linear):
             torch.nn.init.xavier_uniform_(m.weight)
             m.bias.data.fill_(0.01)
