+++ net2brain/architectures/timm_models.py
@@ -1,1352 +1,133 @@
-from pathlib import Path
-from PIL import Image
-from typing import Union, Callable
 
+import warnings
+from .netsetbase import NetSetBase
+from .shared_functions import load_from_json
 import timm
-import torch
+from typing import Callable
 import torch.nn as nn
 from timm.data import resolve_data_config
 from timm.data.constants import IMAGENET_DEFAULT_MEAN, IMAGENET_DEFAULT_STD
 from timm.data.transforms_factory import create_transform as timm_create_transform
 from torchvision import transforms as T
+import torchextractor as tx
+from PIL import Image
 
-MODELS = {
-    "adv_inception_v3": timm.create_model,
-    "cait_m36_384": timm.create_model,
-    "cait_m48_448": timm.create_model,
-    "cait_s24_224": timm.create_model,
-    "cait_s24_384": timm.create_model,
-    "cait_s36_384": timm.create_model,
-    "cait_xs24_384": timm.create_model,
-    "cait_xxs24_224": timm.create_model,
-    "cait_xxs24_384": timm.create_model,
-    "cait_xxs36_224": timm.create_model,
-    "cait_xxs36_384": timm.create_model,
-    "coat_lite_mini": timm.create_model,
-    "coat_lite_small": timm.create_model,
-    "coat_lite_tiny": timm.create_model,
-    "coat_mini": timm.create_model,
-    "coat_tiny": timm.create_model,
-    "convit_base": timm.create_model,
-    "convit_small": timm.create_model,
-    "convit_tiny": timm.create_model,
-    "cspdarknet53": timm.create_model,
-    "cspresnet50": timm.create_model,
-    "cspresnext50": timm.create_model,
-    "deit_base_distilled_patch16_224": timm.create_model,
-    "deit_base_distilled_patch16_384": timm.create_model,
-    "deit_base_patch16_224": timm.create_model,
-    "deit_base_patch16_384": timm.create_model,
-    "deit_small_distilled_patch16_224": timm.create_model,
-    "deit_small_patch16_224": timm.create_model,
-    "deit_tiny_distilled_patch16_224": timm.create_model,
-    "deit_tiny_patch16_224": timm.create_model,
-    "densenet121": timm.create_model,
-    "densenet161": timm.create_model,
-    "densenet169": timm.create_model,
-    "densenet201": timm.create_model,
-    "densenetblur121d": timm.create_model,
-    "dla34": timm.create_model,
-    "dla46_c": timm.create_model,
-    "dla46x_c": timm.create_model,
-    "dla60": timm.create_model,
-    "dla60_res2net": timm.create_model,
-    "dla60_res2next": timm.create_model,
-    "dla60x": timm.create_model,
-    "dla60x_c": timm.create_model,
-    "dla102": timm.create_model,
-    "dla102x": timm.create_model,
-    "dla102x2": timm.create_model,
-    "dla169": timm.create_model,
-    "dm_nfnet_f0": timm.create_model,
-    "dm_nfnet_f1": timm.create_model,
-    "dm_nfnet_f2": timm.create_model,
-    "dm_nfnet_f3": timm.create_model,
-    "dm_nfnet_f4": timm.create_model,
-    "dm_nfnet_f5": timm.create_model,
-    "dm_nfnet_f6": timm.create_model,
-    "dpn68": timm.create_model,
-    "dpn68b": timm.create_model,
-    "dpn92": timm.create_model,
-    "dpn98": timm.create_model,
-    "dpn107": timm.create_model,
-    "dpn131": timm.create_model,
-    "eca_nfnet_l0": timm.create_model,
-    "eca_nfnet_l1": timm.create_model,
-    "eca_nfnet_l2": timm.create_model,
-    "ecaresnet26t": timm.create_model,
-    "ecaresnet50d": timm.create_model,
-    "ecaresnet50d_pruned": timm.create_model,
-    "ecaresnet50t": timm.create_model,
-    "ecaresnet101d": timm.create_model,
-    "ecaresnet101d_pruned": timm.create_model,
-    "ecaresnet269d": timm.create_model,
-    "ecaresnetlight": timm.create_model,
-    "efficientnet_b0": timm.create_model,
-    "efficientnet_b1": timm.create_model,
-    "efficientnet_b1_pruned": timm.create_model,
-    "efficientnet_b2": timm.create_model,
-    "efficientnet_b2_pruned": timm.create_model,
-    "efficientnet_b3": timm.create_model,
-    "efficientnet_b3_pruned": timm.create_model,
-    "efficientnet_b4": timm.create_model,
-    "efficientnet_el": timm.create_model,
-    "efficientnet_el_pruned": timm.create_model,
-    "efficientnet_em": timm.create_model,
-    "efficientnet_es": timm.create_model,
-    "efficientnet_es_pruned": timm.create_model,
-    "efficientnet_lite0": timm.create_model,
-    "efficientnetv2_rw_m": timm.create_model,
-    "efficientnetv2_rw_s": timm.create_model,
-    "ens_adv_inception_resnet_v2": timm.create_model,
-    "ese_vovnet19b_dw": timm.create_model,
-    "ese_vovnet39b": timm.create_model,
-    "fbnetc_100": timm.create_model,
-    "gernet_l": timm.create_model,
-    "gernet_m": timm.create_model,
-    "gernet_s": timm.create_model,
-    "ghostnet_100": timm.create_model,
-    "gluon_inception_v3": timm.create_model,
-    "gluon_resnet18_v1b": timm.create_model,
-    "gluon_resnet34_v1b": timm.create_model,
-    "gluon_resnet50_v1b": timm.create_model,
-    "gluon_resnet50_v1c": timm.create_model,
-    "gluon_resnet50_v1d": timm.create_model,
-    "gluon_resnet50_v1s": timm.create_model,
-    "gluon_resnet101_v1b": timm.create_model,
-    "gluon_resnet101_v1c": timm.create_model,
-    "gluon_resnet101_v1d": timm.create_model,
-    "gluon_resnet101_v1s": timm.create_model,
-    "gluon_resnet152_v1b": timm.create_model,
-    "gluon_resnet152_v1c": timm.create_model,
-    "gluon_resnet152_v1d": timm.create_model,
-    "gluon_resnet152_v1s": timm.create_model,
-    "gluon_resnext50_32x4d": timm.create_model,
-    "gluon_resnext101_32x4d": timm.create_model,
-    "gluon_resnext101_64x4d": timm.create_model,
-    "gluon_senet154": timm.create_model,
-    "gluon_seresnext50_32x4d": timm.create_model,
-    "gluon_seresnext101_32x4d": timm.create_model,
-    "gluon_seresnext101_64x4d": timm.create_model,
-    "gluon_xception65": timm.create_model,
-    "gmixer_24_224": timm.create_model,
-    "gmlp_s16_224": timm.create_model,
-    "hardcorenas_a": timm.create_model,
-    "hardcorenas_b": timm.create_model,
-    "hardcorenas_c": timm.create_model,
-    "hardcorenas_d": timm.create_model,
-    "hardcorenas_e": timm.create_model,
-    "hardcorenas_f": timm.create_model,
-    "hrnet_w18": timm.create_model,
-    "hrnet_w18_small": timm.create_model,
-    "hrnet_w18_small_v2": timm.create_model,
-    "hrnet_w30": timm.create_model,
-    "hrnet_w32": timm.create_model,
-    "hrnet_w40": timm.create_model,
-    "hrnet_w44": timm.create_model,
-    "hrnet_w48": timm.create_model,
-    "hrnet_w64": timm.create_model,
-    "ig_resnext101_32x8d": timm.create_model,
-    "ig_resnext101_32x16d": timm.create_model,
-    "ig_resnext101_32x32d": timm.create_model,
-    "ig_resnext101_32x48d": timm.create_model,
-    "inception_resnet_v2": timm.create_model,
-    "inception_v3": timm.create_model,
-    "inception_v4": timm.create_model,
-    "legacy_senet154": timm.create_model,
-    "legacy_seresnet18": timm.create_model,
-    "legacy_seresnet34": timm.create_model,
-    "legacy_seresnet50": timm.create_model,
-    "legacy_seresnet101": timm.create_model,
-    "legacy_seresnet152": timm.create_model,
-    "legacy_seresnext26_32x4d": timm.create_model,
-    "legacy_seresnext50_32x4d": timm.create_model,
-    "legacy_seresnext101_32x4d": timm.create_model,
-    "levit_384": timm.create_model,
-    "mixer_b16_224": timm.create_model,
-    "mixer_b16_224_in21k": timm.create_model,
-    "mixer_b16_224_miil": timm.create_model,
-    "mixer_b16_224_miil_in21k": timm.create_model,
-    "mixer_l16_224": timm.create_model,
-    "mixer_l16_224_in21k": timm.create_model,
-    "mixnet_l": timm.create_model,
-    "mixnet_m": timm.create_model,
-    "mixnet_s": timm.create_model,
-    "mixnet_xl": timm.create_model,
-    "mnasnet_100": timm.create_model,
-    "mobilenetv2_100": timm.create_model,
-    "mobilenetv2_110d": timm.create_model,
-    "mobilenetv2_120d": timm.create_model,
-    "mobilenetv2_140": timm.create_model,
-    "mobilenetv3_large_100": timm.create_model,
-    "mobilenetv3_large_100_miil": timm.create_model,
-    "mobilenetv3_large_100_miil_in21k": timm.create_model,
-    "mobilenetv3_rw": timm.create_model,
-    "nasnetalarge": timm.create_model,
-    "nf_regnet_b1": timm.create_model,
-    "nf_resnet50": timm.create_model,
-    "nfnet_l0": timm.create_model,
-    "pit_b_224": timm.create_model,
-    "pit_b_distilled_224": timm.create_model,
-    "pit_s_224": timm.create_model,
-    "pit_s_distilled_224": timm.create_model,
-    "pit_ti_224": timm.create_model,
-    "pit_ti_distilled_224": timm.create_model,
-    "pit_xs_224": timm.create_model,
-    "pit_xs_distilled_224": timm.create_model,
-    "pnasnet5large": timm.create_model,
-    "regnetx_002": timm.create_model,
-    "regnetx_004": timm.create_model,
-    "regnetx_006": timm.create_model,
-    "regnetx_008": timm.create_model,
-    "regnetx_016": timm.create_model,
-    "regnetx_032": timm.create_model,
-    "regnetx_040": timm.create_model,
-    "regnetx_064": timm.create_model,
-    "regnetx_080": timm.create_model,
-    "regnetx_120": timm.create_model,
-    "regnetx_160": timm.create_model,
-    "regnetx_320": timm.create_model,
-    "regnety_002": timm.create_model,
-    "regnety_004": timm.create_model,
-    "regnety_006": timm.create_model,
-    "regnety_008": timm.create_model,
-    "regnety_016": timm.create_model,
-    "regnety_032": timm.create_model,
-    "regnety_040": timm.create_model,
-    "regnety_064": timm.create_model,
-    "regnety_080": timm.create_model,
-    "regnety_120": timm.create_model,
-    "regnety_160": timm.create_model,
-    "regnety_320": timm.create_model,
-    "repvgg_a2": timm.create_model,
-    "repvgg_b0": timm.create_model,
-    "repvgg_b1": timm.create_model,
-    "repvgg_b1g4": timm.create_model,
-    "repvgg_b2": timm.create_model,
-    "repvgg_b2g4": timm.create_model,
-    "repvgg_b3": timm.create_model,
-    "repvgg_b3g4": timm.create_model,
-    "res2net50_14w_8s": timm.create_model,
-    "res2net50_26w_4s": timm.create_model,
-    "res2net50_26w_6s": timm.create_model,
-    "res2net50_26w_8s": timm.create_model,
-    "res2net50_48w_2s": timm.create_model,
-    "res2net101_26w_4s": timm.create_model,
-    "res2next50": timm.create_model,
-    "resmlp_12_224": timm.create_model,
-    "resmlp_12_distilled_224": timm.create_model,
-    "resmlp_24_224": timm.create_model,
-    "resmlp_24_distilled_224": timm.create_model,
-    "resmlp_36_224": timm.create_model,
-    "resmlp_36_distilled_224": timm.create_model,
-    "resmlp_big_24_224": timm.create_model,
-    "resmlp_big_24_224_in22ft1k": timm.create_model,
-    "resmlp_big_24_distilled_224": timm.create_model,
-    "resnet18": timm.create_model,
-    "resnet18d": timm.create_model,
-    "resnet26": timm.create_model,
-    "resnet26d": timm.create_model,
-    "resnet34": timm.create_model,
-    "resnet34d": timm.create_model,
-    "resnet50": timm.create_model,
-    "resnet50d": timm.create_model,
-    "resnet51q": timm.create_model,
-    "resnet101d": timm.create_model,
-    "resnet152d": timm.create_model,
-    "resnet200d": timm.create_model,
-    "resnetblur50": timm.create_model,
-    "resnetrs50": timm.create_model,
-    "resnetrs101": timm.create_model,
-    "resnetrs152": timm.create_model,
-    "resnetrs200": timm.create_model,
-    "resnetrs270": timm.create_model,
-    "resnetrs350": timm.create_model,
-    "resnetrs420": timm.create_model,
-    "resnetv2_50x1_bit_distilled": timm.create_model,
-    "resnetv2_50x1_bitm": timm.create_model,
-    "resnetv2_50x1_bitm_in21k": timm.create_model,
-    "resnetv2_50x3_bitm": timm.create_model,
-    "resnetv2_50x3_bitm_in21k": timm.create_model,
-    "resnetv2_101x1_bitm": timm.create_model,
-    "resnetv2_101x1_bitm_in21k": timm.create_model,
-    "resnetv2_101x3_bitm": timm.create_model,
-    "resnetv2_101x3_bitm_in21k": timm.create_model,
-    "resnetv2_152x2_bit_teacher": timm.create_model,
-    "resnetv2_152x2_bit_teacher_384": timm.create_model,
-    "resnetv2_152x2_bitm": timm.create_model,
-    "resnetv2_152x2_bitm_in21k": timm.create_model,
-    "resnetv2_152x4_bitm": timm.create_model,
-    "resnetv2_152x4_bitm_in21k": timm.create_model,
-    "resnext50_32x4d": timm.create_model,
-    "resnext50d_32x4d": timm.create_model,
-    "resnext101_32x8d": timm.create_model,
-    "selecsls42b": timm.create_model,
-    "selecsls60": timm.create_model,
-    "selecsls60b": timm.create_model,
-    "semnasnet_100": timm.create_model,
-    "seresnet50": timm.create_model,
-    "seresnet152d": timm.create_model,
-    "seresnext26d_32x4d": timm.create_model,
-    "seresnext26t_32x4d": timm.create_model,
-    "seresnext50_32x4d": timm.create_model,
-    "spnasnet_100": timm.create_model,
-    "ssl_resnet18": timm.create_model,
-    "ssl_resnet50": timm.create_model,
-    "ssl_resnext50_32x4d": timm.create_model,
-    "ssl_resnext101_32x4d": timm.create_model,
-    "ssl_resnext101_32x8d": timm.create_model,
-    "ssl_resnext101_32x16d": timm.create_model,
-    "swin_base_patch4_window7_224": timm.create_model,
-    "swin_base_patch4_window7_224_in22k": timm.create_model,
-    "swin_base_patch4_window12_384": timm.create_model,
-    "swin_base_patch4_window12_384_in22k": timm.create_model,
-    "swin_large_patch4_window7_224": timm.create_model,
-    "swin_large_patch4_window7_224_in22k": timm.create_model,
-    "swin_large_patch4_window12_384": timm.create_model,
-    "swin_large_patch4_window12_384_in22k": timm.create_model,
-    "swin_small_patch4_window7_224": timm.create_model,
-    "swin_tiny_patch4_window7_224": timm.create_model,
-    "swsl_resnet18": timm.create_model,
-    "swsl_resnet50": timm.create_model,
-    "swsl_resnext50_32x4d": timm.create_model,
-    "swsl_resnext101_32x4d": timm.create_model,
-    "swsl_resnext101_32x8d": timm.create_model,
-    "swsl_resnext101_32x16d": timm.create_model,
-    "tf_efficientnet_b0": timm.create_model,
-    "tf_efficientnet_b0_ap": timm.create_model,
-    "tf_efficientnet_b0_ns": timm.create_model,
-    "tf_efficientnet_b1": timm.create_model,
-    "tf_efficientnet_b1_ap": timm.create_model,
-    "tf_efficientnet_b1_ns": timm.create_model,
-    "tf_efficientnet_b2": timm.create_model,
-    "tf_efficientnet_b2_ap": timm.create_model,
-    "tf_efficientnet_b2_ns": timm.create_model,
-    "tf_efficientnet_b3": timm.create_model,
-    "tf_efficientnet_b3_ap": timm.create_model,
-    "tf_efficientnet_b3_ns": timm.create_model,
-    "tf_efficientnet_b4": timm.create_model,
-    "tf_efficientnet_b4_ap": timm.create_model,
-    "tf_efficientnet_b4_ns": timm.create_model,
-    "tf_efficientnet_b5": timm.create_model,
-    "tf_efficientnet_b5_ap": timm.create_model,
-    "tf_efficientnet_b5_ns": timm.create_model,
-    "tf_efficientnet_b6": timm.create_model,
-    "tf_efficientnet_b6_ap": timm.create_model,
-    "tf_efficientnet_b6_ns": timm.create_model,
-    "tf_efficientnet_b7": timm.create_model,
-    "tf_efficientnet_b7_ap": timm.create_model,
-    "tf_efficientnet_b7_ns": timm.create_model,
-    "tf_efficientnet_b8": timm.create_model,
-    "tf_efficientnet_b8_ap": timm.create_model,
-    "tf_efficientnet_cc_b0_4e": timm.create_model,
-    "tf_efficientnet_cc_b0_8e": timm.create_model,
-    "tf_efficientnet_cc_b1_8e": timm.create_model,
-    "tf_efficientnet_el": timm.create_model,
-    "tf_efficientnet_em": timm.create_model,
-    "tf_efficientnet_es": timm.create_model,
-    "tf_efficientnet_l2_ns": timm.create_model,
-    "tf_efficientnet_l2_ns_475": timm.create_model,
-    "tf_efficientnet_lite0": timm.create_model,
-    "tf_efficientnet_lite1": timm.create_model,
-    "tf_efficientnet_lite2": timm.create_model,
-    "tf_efficientnet_lite3": timm.create_model,
-    "tf_efficientnet_lite4": timm.create_model,
-    "tf_efficientnetv2_b0": timm.create_model,
-    "tf_efficientnetv2_b1": timm.create_model,
-    "tf_efficientnetv2_b2": timm.create_model,
-    "tf_efficientnetv2_b3": timm.create_model,
-    "tf_efficientnetv2_l": timm.create_model,
-    "tf_efficientnetv2_l_in21ft1k": timm.create_model,
-    "tf_efficientnetv2_l_in21k": timm.create_model,
-    "tf_efficientnetv2_m": timm.create_model,
-    "tf_efficientnetv2_m_in21ft1k": timm.create_model,
-    "tf_efficientnetv2_m_in21k": timm.create_model,
-    "tf_efficientnetv2_s": timm.create_model,
-    "tf_efficientnetv2_s_in21ft1k": timm.create_model,
-    "tf_efficientnetv2_s_in21k": timm.create_model,
-    "tf_inception_v3": timm.create_model,
-    "tf_mixnet_l": timm.create_model,
-    "tf_mixnet_m": timm.create_model,
-    "tf_mixnet_s": timm.create_model,
-    "tf_mobilenetv3_large_075": timm.create_model,
-    "tf_mobilenetv3_large_100": timm.create_model,
-    "tf_mobilenetv3_large_minimal_100": timm.create_model,
-    "tf_mobilenetv3_small_075": timm.create_model,
-    "tf_mobilenetv3_small_100": timm.create_model,
-    "tf_mobilenetv3_small_minimal_100": timm.create_model,
-    "tnt_s_patch16_224": timm.create_model,
-    "tv_densenet121": timm.create_model,
-    "tv_resnet34": timm.create_model,
-    "tv_resnet50": timm.create_model,
-    "tv_resnet101": timm.create_model,
-    "tv_resnet152": timm.create_model,
-    "tv_resnext50_32x4d": timm.create_model,
-    "twins_pcpvt_base": timm.create_model,
-    "twins_pcpvt_large": timm.create_model,
-    "twins_pcpvt_small": timm.create_model,
-    "twins_svt_base": timm.create_model,
-    "twins_svt_large": timm.create_model,
-    "twins_svt_small": timm.create_model,
-    "vgg11": timm.create_model,
-    "vgg11_bn": timm.create_model,
-    "vgg13": timm.create_model,
-    "vgg13_bn": timm.create_model,
-    "vgg16": timm.create_model,
-    "vgg16_bn": timm.create_model,
-    "vgg19": timm.create_model,
-    "vgg19_bn": timm.create_model,
-    "visformer_small": timm.create_model,
-    "vit_base_patch16_224": timm.create_model,
-    "vit_base_patch16_224_in21k": timm.create_model,
-    "vit_base_patch16_224_miil": timm.create_model,
-    "vit_base_patch16_224_miil_in21k": timm.create_model,
-    "vit_base_patch16_384": timm.create_model,
-    "vit_base_patch32_224": timm.create_model,
-    "vit_base_patch32_224_in21k": timm.create_model,
-    "vit_base_patch32_384": timm.create_model,
-    "vit_base_r50_s16_224_in21k": timm.create_model,
-    "vit_base_r50_s16_384": timm.create_model,
-    "vit_huge_patch14_224_in21k": timm.create_model,
-    "vit_large_patch16_224": timm.create_model,
-    "vit_large_patch16_224_in21k": timm.create_model,
-    "vit_large_patch16_384": timm.create_model,
-    "vit_large_patch32_224_in21k": timm.create_model,
-    "vit_large_patch32_384": timm.create_model,
-    "vit_large_r50_s32_224": timm.create_model,
-    "vit_large_r50_s32_224_in21k": timm.create_model,
-    "vit_large_r50_s32_384": timm.create_model,
-    "vit_small_patch16_224": timm.create_model,
-    "vit_small_patch16_224_in21k": timm.create_model,
-    "vit_small_patch16_384": timm.create_model,
-    "vit_small_patch32_224": timm.create_model,
-    "vit_small_patch32_224_in21k": timm.create_model,
-    "vit_small_patch32_384": timm.create_model,
-    "vit_small_r26_s32_224": timm.create_model,
-    "vit_small_r26_s32_224_in21k": timm.create_model,
-    "vit_small_r26_s32_384": timm.create_model,
-    "vit_tiny_patch16_224": timm.create_model,
-    "vit_tiny_patch16_224_in21k": timm.create_model,
-    "vit_tiny_patch16_384": timm.create_model,
-    "vit_tiny_r_s16_p8_224": timm.create_model,
-    "vit_tiny_r_s16_p8_224_in21k": timm.create_model,
-    "vit_tiny_r_s16_p8_384": timm.create_model,
-    "wide_resnet50_2": timm.create_model,
-    "wide_resnet101_2": timm.create_model,
-    "xception": timm.create_model,
-    "xception41": timm.create_model,
-    "xception65": timm.create_model,
-    "xception71": timm.create_model,
-}
+class Timm(NetSetBase):
 
-MODEL_NODES = {
-    "adv_inception_v3": [],
-    "cait_m36_384": [],
-    "cait_m48_448": [],
-    "cait_s24_224": [],
-    "cait_s24_384": [],
-    "cait_s36_384": [],
-    "cait_xs24_384": [],
-    "cait_xxs24_224": [],
-    "cait_xxs24_384": [],
-    "cait_xxs36_224": [],
-    "cait_xxs36_384": [],
-    "coat_lite_mini": [],
-    "coat_lite_small": [],
-    "coat_lite_tiny": [],
-    "coat_mini": [],
-    "coat_tiny": [],
-    "convit_base": [],
-    "convit_small": [],
-    "convit_tiny": [],
-    "cspdarknet53": [],
-    "cspresnet50": [],
-    "cspresnext50": [],
-    "deit_base_distilled_patch16_224": [],
-    "deit_base_distilled_patch16_384": [],
-    "deit_base_patch16_224": [],
-    "deit_base_patch16_384": [],
-    "deit_small_distilled_patch16_224": [],
-    "deit_small_patch16_224": [],
-    "deit_tiny_distilled_patch16_224": [],
-    "deit_tiny_patch16_224": [],
-    "densenet121": [],
-    "densenet161": [],
-    "densenet169": [],
-    "densenet201": [],
-    "densenetblur121d": [],
-    "dla34": [],
-    "dla46_c": [],
-    "dla46x_c": [],
-    "dla60": [],
-    "dla60_res2net": [],
-    "dla60_res2next": [],
-    "dla60x": [],
-    "dla60x_c": [],
-    "dla102": [],
-    "dla102x": [],
-    "dla102x2": [],
-    "dla169": [],
-    "dm_nfnet_f0": [],
-    "dm_nfnet_f1": [],
-    "dm_nfnet_f2": [],
-    "dm_nfnet_f3": [],
-    "dm_nfnet_f4": [],
-    "dm_nfnet_f5": [],
-    "dm_nfnet_f6": [],
-    "dpn68": [],
-    "dpn68b": [],
-    "dpn92": [],
-    "dpn98": [],
-    "dpn107": [],
-    "dpn131": [],
-    "eca_nfnet_l0": [],
-    "eca_nfnet_l1": [],
-    "eca_nfnet_l2": [],
-    "ecaresnet26t": [],
-    "ecaresnet50d": [],
-    "ecaresnet50d_pruned": [],
-    "ecaresnet50t": [],
-    "ecaresnet101d": [],
-    "ecaresnet101d_pruned": [],
-    "ecaresnet269d": [],
-    "ecaresnetlight": [],
-    "efficientnet_b0": [],
-    "efficientnet_b1": [],
-    "efficientnet_b1_pruned": [],
-    "efficientnet_b2": [],
-    "efficientnet_b2_pruned": [],
-    "efficientnet_b3": [],
-    "efficientnet_b3_pruned": [],
-    "efficientnet_b4": [],
-    "efficientnet_el": [],
-    "efficientnet_el_pruned": [],
-    "efficientnet_em": [],
-    "efficientnet_es": [],
-    "efficientnet_es_pruned": [],
-    "efficientnet_lite0": [],
-    "efficientnetv2_rw_m": [],
-    "efficientnetv2_rw_s": [],
-    "ens_adv_inception_resnet_v2": [],
-    "ese_vovnet19b_dw": [],
-    "ese_vovnet39b": [],
-    "fbnetc_100": [],
-    "gernet_l": [],
-    "gernet_m": [],
-    "gernet_s": [],
-    "ghostnet_100": [],
-    "gluon_inception_v3": [],
-    "gluon_resnet18_v1b": [],
-    "gluon_resnet34_v1b": [],
-    "gluon_resnet50_v1b": [],
-    "gluon_resnet50_v1c": [],
-    "gluon_resnet50_v1d": [],
-    "gluon_resnet50_v1s": [],
-    "gluon_resnet101_v1b": [],
-    "gluon_resnet101_v1c": [],
-    "gluon_resnet101_v1d": [],
-    "gluon_resnet101_v1s": [],
-    "gluon_resnet152_v1b": [],
-    "gluon_resnet152_v1c": [],
-    "gluon_resnet152_v1d": [],
-    "gluon_resnet152_v1s": [],
-    "gluon_resnext50_32x4d": [],
-    "gluon_resnext101_32x4d": [],
-    "gluon_resnext101_64x4d": [],
-    "gluon_senet154": [],
-    "gluon_seresnext50_32x4d": [],
-    "gluon_seresnext101_32x4d": [],
-    "gluon_seresnext101_64x4d": [],
-    "gluon_xception65": [],
-    "gmixer_24_224": [],
-    "gmlp_s16_224": [],
-    "hardcorenas_a": [],
-    "hardcorenas_b": [],
-    "hardcorenas_c": [],
-    "hardcorenas_d": [],
-    "hardcorenas_e": [],
-    "hardcorenas_f": [],
-    "hrnet_w18": [],
-    "hrnet_w18_small": [],
-    "hrnet_w18_small_v2": [],
-    "hrnet_w30": [],
-    "hrnet_w32": [],
-    "hrnet_w40": [],
-    "hrnet_w44": [],
-    "hrnet_w48": [],
-    "hrnet_w64": [],
-    "ig_resnext101_32x8d": [],
-    "ig_resnext101_32x16d": [],
-    "ig_resnext101_32x32d": [],
-    "ig_resnext101_32x48d": [],
-    "inception_resnet_v2": [],
-    "inception_v3": [],
-    "inception_v4": [],
-    "legacy_senet154": [],
-    "legacy_seresnet18": [],
-    "legacy_seresnet34": [],
-    "legacy_seresnet50": [],
-    "legacy_seresnet101": [],
-    "legacy_seresnet152": [],
-    "legacy_seresnext26_32x4d": [],
-    "legacy_seresnext50_32x4d": [],
-    "legacy_seresnext101_32x4d": [],
-    "levit_384": [],
-    "mixer_b16_224": [],
-    "mixer_b16_224_in21k": [],
-    "mixer_b16_224_miil": [],
-    "mixer_b16_224_miil_in21k": [],
-    "mixer_l16_224": [],
-    "mixer_l16_224_in21k": [],
-    "mixnet_l": [],
-    "mixnet_m": [],
-    "mixnet_s": [],
-    "mixnet_xl": [],
-    "mnasnet_100": [],
-    "mobilenetv2_100": [],
-    "mobilenetv2_110d": [],
-    "mobilenetv2_120d": [],
-    "mobilenetv2_140": [],
-    "mobilenetv3_large_100": [],
-    "mobilenetv3_large_100_miil": [],
-    "mobilenetv3_large_100_miil_in21k": [],
-    "mobilenetv3_rw": [],
-    "nasnetalarge": [],
-    "nf_regnet_b1": [],
-    "nf_resnet50": [],
-    "nfnet_l0": [],
-    "pit_b_224": [],
-    "pit_b_distilled_224": [],
-    "pit_s_224": [],
-    "pit_s_distilled_224": [],
-    "pit_ti_224": [],
-    "pit_ti_distilled_224": [],
-    "pit_xs_224": [],
-    "pit_xs_distilled_224": [],
-    "pnasnet5large": [],
-    "regnetx_002": [],
-    "regnetx_004": [],
-    "regnetx_006": [],
-    "regnetx_008": [],
-    "regnetx_016": [],
-    "regnetx_032": [],
-    "regnetx_040": [],
-    "regnetx_064": [],
-    "regnetx_080": [],
-    "regnetx_120": [],
-    "regnetx_160": [],
-    "regnetx_320": [],
-    "regnety_002": [],
-    "regnety_004": [],
-    "regnety_006": [],
-    "regnety_008": [],
-    "regnety_016": [],
-    "regnety_032": [],
-    "regnety_040": [],
-    "regnety_064": [],
-    "regnety_080": [],
-    "regnety_120": [],
-    "regnety_160": [],
-    "regnety_320": [],
-    "repvgg_a2": [],
-    "repvgg_b0": [],
-    "repvgg_b1": [],
-    "repvgg_b1g4": [],
-    "repvgg_b2": [],
-    "repvgg_b2g4": [],
-    "repvgg_b3": [],
-    "repvgg_b3g4": [],
-    "res2net50_14w_8s": [],
-    "res2net50_26w_4s": [],
-    "res2net50_26w_6s": [],
-    "res2net50_26w_8s": [],
-    "res2net50_48w_2s": [],
-    "res2net101_26w_4s": [],
-    "res2next50": [],
-    "resmlp_12_224": [],
-    "resmlp_12_distilled_224": [],
-    "resmlp_24_224": [],
-    "resmlp_24_distilled_224": [],
-    "resmlp_36_224": [],
-    "resmlp_36_distilled_224": [],
-    "resmlp_big_24_224": [],
-    "resmlp_big_24_224_in22ft1k": [],
-    "resmlp_big_24_distilled_224": [],
-    "resnet18": [],
-    "resnet18d": [],
-    "resnet26": [],
-    "resnet26d": [],
-    "resnet34": [],
-    "resnet34d": [],
-    "resnet50": [],
-    "resnet50d": [],
-    "resnet51q": [],
-    "resnet101d": [],
-    "resnet152d": [],
-    "resnet200d": [],
-    "resnetblur50": [],
-    "resnetrs50": [],
-    "resnetrs101": [],
-    "resnetrs152": [],
-    "resnetrs200": [],
-    "resnetrs270": [],
-    "resnetrs350": [],
-    "resnetrs420": [],
-    "resnetv2_50x1_bit_distilled": [],
-    "resnetv2_50x1_bitm": [],
-    "resnetv2_50x1_bitm_in21k": [],
-    "resnetv2_50x3_bitm": [],
-    "resnetv2_50x3_bitm_in21k": [],
-    "resnetv2_101x1_bitm": [],
-    "resnetv2_101x1_bitm_in21k": [],
-    "resnetv2_101x3_bitm": [],
-    "resnetv2_101x3_bitm_in21k": [],
-    "resnetv2_152x2_bit_teacher": [],
-    "resnetv2_152x2_bit_teacher_384": [],
-    "resnetv2_152x2_bitm": [],
-    "resnetv2_152x2_bitm_in21k": [],
-    "resnetv2_152x4_bitm": [],
-    "resnetv2_152x4_bitm_in21k": [],
-    "resnext50_32x4d": [],
-    "resnext50d_32x4d": [],
-    "resnext101_32x8d": [],
-    "selecsls42b": [],
-    "selecsls60": [],
-    "selecsls60b": [],
-    "semnasnet_100": [],
-    "seresnet50": [],
-    "seresnet152d": [],
-    "seresnext26d_32x4d": [],
-    "seresnext26t_32x4d": [],
-    "seresnext50_32x4d": [],
-    "spnasnet_100": [],
-    "ssl_resnet18": [],
-    "ssl_resnet50": [],
-    "ssl_resnext50_32x4d": [],
-    "ssl_resnext101_32x4d": [],
-    "ssl_resnext101_32x8d": [],
-    "ssl_resnext101_32x16d": [],
-    "swin_base_patch4_window7_224": [],
-    "swin_base_patch4_window7_224_in22k": [],
-    "swin_base_patch4_window12_384": [],
-    "swin_base_patch4_window12_384_in22k": [],
-    "swin_large_patch4_window7_224": [],
-    "swin_large_patch4_window7_224_in22k": [],
-    "swin_large_patch4_window12_384": [],
-    "swin_large_patch4_window12_384_in22k": [],
-    "swin_small_patch4_window7_224": [],
-    "swin_tiny_patch4_window7_224": [],
-    "swsl_resnet18": [],
-    "swsl_resnet50": [],
-    "swsl_resnext50_32x4d": [],
-    "swsl_resnext101_32x4d": [],
-    "swsl_resnext101_32x8d": [],
-    "swsl_resnext101_32x16d": [],
-    "tf_efficientnet_b0": [],
-    "tf_efficientnet_b0_ap": [],
-    "tf_efficientnet_b0_ns": [],
-    "tf_efficientnet_b1": [],
-    "tf_efficientnet_b1_ap": [],
-    "tf_efficientnet_b1_ns": [],
-    "tf_efficientnet_b2": [],
-    "tf_efficientnet_b2_ap": [],
-    "tf_efficientnet_b2_ns": [],
-    "tf_efficientnet_b3": [],
-    "tf_efficientnet_b3_ap": [],
-    "tf_efficientnet_b3_ns": [],
-    "tf_efficientnet_b4": [],
-    "tf_efficientnet_b4_ap": [],
-    "tf_efficientnet_b4_ns": [],
-    "tf_efficientnet_b5": [],
-    "tf_efficientnet_b5_ap": [],
-    "tf_efficientnet_b5_ns": [],
-    "tf_efficientnet_b6": [],
-    "tf_efficientnet_b6_ap": [],
-    "tf_efficientnet_b6_ns": [],
-    "tf_efficientnet_b7": [],
-    "tf_efficientnet_b7_ap": [],
-    "tf_efficientnet_b7_ns": [],
-    "tf_efficientnet_b8": [],
-    "tf_efficientnet_b8_ap": [],
-    "tf_efficientnet_cc_b0_4e": [],
-    "tf_efficientnet_cc_b0_8e": [],
-    "tf_efficientnet_cc_b1_8e": [],
-    "tf_efficientnet_el": [],
-    "tf_efficientnet_em": [],
-    "tf_efficientnet_es": [],
-    "tf_efficientnet_l2_ns": [],
-    "tf_efficientnet_l2_ns_475": [],
-    "tf_efficientnet_lite0": [],
-    "tf_efficientnet_lite1": [],
-    "tf_efficientnet_lite2": [],
-    "tf_efficientnet_lite3": [],
-    "tf_efficientnet_lite4": [],
-    "tf_efficientnetv2_b0": [],
-    "tf_efficientnetv2_b1": [],
-    "tf_efficientnetv2_b2": [],
-    "tf_efficientnetv2_b3": [],
-    "tf_efficientnetv2_l": [],
-    "tf_efficientnetv2_l_in21ft1k": [],
-    "tf_efficientnetv2_l_in21k": [],
-    "tf_efficientnetv2_m": [],
-    "tf_efficientnetv2_m_in21ft1k": [],
-    "tf_efficientnetv2_m_in21k": [],
-    "tf_efficientnetv2_s": [],
-    "tf_efficientnetv2_s_in21ft1k": [],
-    "tf_efficientnetv2_s_in21k": [],
-    "tf_inception_v3": [],
-    "tf_mixnet_l": [],
-    "tf_mixnet_m": [],
-    "tf_mixnet_s": [],
-    "tf_mobilenetv3_large_075": [],
-    "tf_mobilenetv3_large_100": [],
-    "tf_mobilenetv3_large_minimal_100": [],
-    "tf_mobilenetv3_small_075": [],
-    "tf_mobilenetv3_small_100": [],
-    "tf_mobilenetv3_small_minimal_100": [],
-    "tnt_s_patch16_224": [],
-    "tv_densenet121": [],
-    "tv_resnet34": [],
-    "tv_resnet50": [],
-    "tv_resnet101": [],
-    "tv_resnet152": [],
-    "tv_resnext50_32x4d": [],
-    "twins_pcpvt_base": [],
-    "twins_pcpvt_large": [],
-    "twins_pcpvt_small": [],
-    "twins_svt_base": [],
-    "twins_svt_large": [],
-    "twins_svt_small": [],
-    "vgg11": [],
-    "vgg11_bn": [],
-    "vgg13": [],
-    "vgg13_bn": [],
-    "vgg16": [],
-    "vgg16_bn": [],
-    "vgg19": [],
-    "vgg19_bn": [],
-    "visformer_small": [],
-    "vit_base_patch16_224": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_base_patch16_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_base_patch16_224_miil": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_base_patch16_224_miil_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_base_patch16_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_base_patch32_224": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_base_patch32_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_base_patch32_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_base_r50_s16_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_base_r50_s16_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_huge_patch14_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_large_patch16_224": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_large_patch16_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_large_patch16_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_large_patch32_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_large_patch32_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_large_r50_s32_224": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_large_r50_s32_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_large_r50_s32_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_small_patch16_224": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_small_patch16_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_small_patch16_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_small_patch32_224": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_small_patch32_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_small_patch32_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_small_r26_s32_224": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_small_r26_s32_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_small_r26_s32_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_tiny_patch16_224": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_tiny_patch16_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_tiny_patch16_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_tiny_r_s16_p8_224": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_tiny_r_s16_p8_224_in21k": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "vit_tiny_r_s16_p8_384": [
-        "blocks.0",
-        "blocks.1",
-        "blocks.2",
-        "blocks.3",
-        "blocks.4",
-        "blocks.5",
-        "blocks.6",
-        "blocks.7",
-        "blocks.8",
-        "blocks.9",
-        "blocks.10",
-        "blocks.11",
-    ],
-    "wide_resnet50_2": [],
-    "wide_resnet101_2": [],
-    "xception": [],
-    "xception41": [],
-    "xception65": [],
-    "xception71": [],
-}
+    def __init__(self, model_name, device):
+        self.supported_data_types = ['image', 'video']
+        self.netset_name = "timm"
+        self.model_name = model_name
+        self.device = device
+        self.config_path = "net2brain/architectures/configs/timm.json"
 
 
-def create_transform(model: nn.Module) -> T.Compose:
-    """
-    Creates a evaluation transform for the given TIMM model.
-    """
-    config = resolve_data_config({}, model=model, verbose=True)
-    config = {
-        "input_size": config.pop("input_size", (3, 224, 224)),
-        "mean": config.pop("mean", IMAGENET_DEFAULT_MEAN),
-        "std": config.pop("std", IMAGENET_DEFAULT_STD),
-        "crop_pct": config.pop("crop_pct", 1.0),
-        "interpolation": config.pop("interpolation", "bicubic"),
-    }
-    return timm_create_transform(**config, is_training=False)
+    def get_preprocessing_function(self, data_type):
+        if data_type == 'image':
+            return self.image_preprocessing
+        elif data_type == 'video':
+            warnings.warn("Models only support image-data. Will average video frames")
+            return self.video_preprocessing
+        else:
+            raise ValueError(f"Unsupported data type for {self.netset_name}: {data_type}")
+        
+
+    def get_feature_cleaner(self, data_type):
+        if data_type == 'image':
+            return Timm.clean_extracted_features
+        elif data_type == 'video':
+            return Timm.clean_extracted_features
+        else:
+            raise ValueError(f"Unsupported data type for {self.netset_name}: {data_type}")
+        
+
+    def get_model(self, pretrained):
+
+        # Load attributes from the json
+        model_attributes = load_from_json(self.config_path, self.model_name)
+
+        # Set the layers and model function from the attributes
+        self.layers = model_attributes["nodes"]
+    
+        try:
+            self.loaded_model = model_attributes["model_function"](self.model_name, pretrained=pretrained, features_only=True)
+        except:
+            self.loaded_model = model_attributes["model_function"](self.model_name, pretrained=pretrained)
+
+        # Model to device
+        self.loaded_model.to(self.device)
+
+        # Randomize weights
+        if not pretrained:
+            self.loaded_model.apply(self.randomize_weights)
+
+        # Put in eval mode
+        self.loaded_model.eval()
+
+    
+    def image_preprocessing(self, image, model_name, device):
+        preprocesser = self.create_preprocess()
+        return preprocesser(image, model_name, device)
+    
+    def video_preprocessing(self, frame, model_name, device):
+        preprocesser = self.create_preprocess()
+        pil_frame = Image.fromarray(frame)
+        return preprocesser(pil_frame, model_name, device)
+    
+    def clean_extracted_features(self, features):
+        return features
+    
+    def extraction_function(self, data, layers_to_extract=None):
+
+        self.layers = self.select_model_layers(layers_to_extract, self.layers, self.loaded_model)
+
+        # Create a extractor instance
+        self.extractor_model = tx.Extractor(self.loaded_model, self.layers)
+
+        # Extract actual features
+        _, features = self.extractor_model(data)
+
+        return features
+
+    def create_transform(self, model: nn.Module) -> T.Compose:
+        """
+        Creates a evaluation transform for the given TIMM model.
+        """
+        config = resolve_data_config({}, model=model, verbose=True)
+        config = {
+            "input_size": config.pop("input_size", (3, 224, 224)),
+            "mean": config.pop("mean", IMAGENET_DEFAULT_MEAN),
+            "std": config.pop("std", IMAGENET_DEFAULT_STD),
+            "crop_pct": config.pop("crop_pct", 1.0),
+            "interpolation": config.pop("interpolation", "bicubic"),
+        }
+        return timm_create_transform(**config, is_training=False)
 
 
-def create_preprocess(model: nn.Module) -> Callable:
-    """
-    Creates a preprocess function for the given TIMM model.
-    """
-    transform = create_transform(model)
-
-    def preprocess(
-        image: Union[str, Path], model_name: str, device: str
-    ) -> torch.Tensor:
-        """Preprocesses image according to the networks needs
-
-        Args:
-            image (str/path): path to image
-            model_name (str): name of the model (sometimes needes to differenciate between model settings)
-
-        Returns:
-            PIL-Image: Preprocesses PIL Image
+    def create_preprocess(self) -> Callable:
+        """
+        Creates a preprocess function for the given TIMM model.
         """
 
-        image = Image.open(image).convert("RGB")
-        image = transform(image).unsqueeze(0).to(device)
-        return image
+        transform = self.create_transform(self.loaded_model)
 
-    return preprocess
+        def preprocess(image, model_name, device):
+            """Preprocesses image according to the networks needs
+
+            Args:
+                image (str/path): path to image
+                model_name (str): name of the model (sometimes needes to differenciate between model settings)
+
+            Returns:
+                PIL-Image: Preprocesses PIL Image
+            """
+
+            image = transform(image).unsqueeze(0).to(device)
+            return image
+
+        return preprocess
+
+
+
+
+
